refactor: process repository
