+++ server/src/services/backup.service.spec.ts
@@ -4,10 +4,9 @@ import { StorageCore } from 'src/cores/storage.core';
 import { ImmichWorker, StorageFolder } from 'src/enum';
 import { IDatabaseRepository } from 'src/interfaces/database.interface';
 import { JobStatus } from 'src/interfaces/job.interface';
-import { IProcessRepository } from 'src/interfaces/process.interface';
 import { IStorageRepository } from 'src/interfaces/storage.interface';
 import { BackupService } from 'src/services/backup.service';
-import { IConfigRepository, ICronRepository, ISystemMetadataRepository } from 'src/types';
+import { IConfigRepository, ICronRepository, IProcessRepository, ISystemMetadataRepository } from 'src/types';
 import { systemConfigStub } from 'test/fixtures/system-config.stub';
 import { mockSpawn, newTestService } from 'test/utils';
 import { describe, Mocked } from 'vitest';
