+++ server/src/types.ts
@@ -14,6 +14,7 @@ import { MemoryRepository } from 'src/repositories/memory.repository';
 import { MetadataRepository } from 'src/repositories/metadata.repository';
 import { NotificationRepository } from 'src/repositories/notification.repository';
 import { OAuthRepository } from 'src/repositories/oauth.repository';
+import { ProcessRepository } from 'src/repositories/process.repository';
 import { ServerInfoRepository } from 'src/repositories/server-info.repository';
 import { SystemMetadataRepository } from 'src/repositories/system-metadata.repository';
 import { MetricGroupRepository, TelemetryRepository } from 'src/repositories/telemetry.repository';
@@ -59,6 +60,7 @@ export type IMetadataRepository = RepositoryInterface<MetadataRepository>;
 export type IMetricGroupRepository = RepositoryInterface<MetricGroupRepository>;
 export type INotificationRepository = RepositoryInterface<NotificationRepository>;
 export type IOAuthRepository = RepositoryInterface<OAuthRepository>;
+export type IProcessRepository = RepositoryInterface<ProcessRepository>;
 export type IServerInfoRepository = RepositoryInterface<ServerInfoRepository>;
 export type ISystemMetadataRepository = RepositoryInterface<SystemMetadataRepository>;
 export type ITelemetryRepository = RepositoryInterface<TelemetryRepository>;
